fix(crawl): 별도 pr 용 코드 원복

+++ src/main/java/com/dekk/card/application/command/CardCreateCommand.java
@@ -11,7 +11,7 @@ public record CardCreateCommand(
     String originId,
     boolean isActive,
     Platform platform,
-    Integer height,
-    Integer weight
+    Double height,
+    Double weight
 ) {
 }
